Rename the header containing the concepts
