+++ test/headers/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_library(memory-headers-tests STATIC EXCLUDE_FROM_ALL
-        allocator.strategy.cpp
         accumulation_buffer.cpp
         any_buffer.cpp
         atomic_pen.cpp
         bitmap.strategy.cpp
         concepts.cpp
+        concepts-polyfill.cpp
         control.cpp
         fixed-pool.pmr.cpp
         holding_pen.cpp
