--- test/headers/allocator.strategy.cpp
@@ -1,1 +1,0 @@
-#include <felspar/memory/allocator.strategy.hpp>
