+++ test/headers/concepts.cpp
@@ -1,1 +1,1 @@
-#include <felspar/concepts.hpp>
+#include <felspar/memory/concepts.hpp>
